Improved date queries and added other options
The date range now accepts `start` and `end` together or an individual
timeframe `day` or `month` with a `range` to indicate how far forward or
backwards to go.

Added the option to flatten the response data into a single array. By
default, it is a 2D array where each nested array indicates a day's
data.